The productName in the pbxproj doesn't seem to get updated. Matching with name instead. Requires Xcode.swift update.
I noticed this today while building another target in the same project
as the original one I’ve been using R.swift with:

If I run `rswift`, I get an error saying that there is not a target
with the name matching the target I’m running, then it prints the
targets name:

```
[R.swift] Target ‘My App Staging’ not found in project file, available
targets are: MyApp, MyAppTests, MyApp
```

Which is obviously not the case because I’m running said target.

I took a look into the `.pbxproj` and found that `My App Staging`’s
`productName` was still set to the `productName` that I had duplicated
it from (My App).

I checked Xcode build settings several times but couldn’t see why this
was the case. It seems that Xcode doesn’t really use this value and
only sets it when creating the project? (I have the build settings
product name set for this Staging target, but it’s not reflected here
by `productName`…).

It is, however, reflected by `name`, so targets should be matched
against name instead.