Reorder Module Imports in Generated Source Code
I updated ImportPrinter to change the order of the imports that are
generated in the output source code. Previously all imports were being
sorted alphabetically and then output to the generated source code. This
was a problem if R.swift.library is being statically linked in another
dynamic framework and the containing framework's module name is
alphabetically after the Rswift module name. In this case, the Swift
compiler does not correctly import the Rswift module.

I changed ImportPrinter to always output modules that are specified
using the --import command line argument before modules that are
extracted from the source code. This should fix the import issue.

I updated CallInformation.swift to pass the list of modules as an array
instead of a set in order to keep the order in which they were specified
on the command line in order to ensure proper ordering of imports in
the generated source code.